requirements.txt
gradio
langchain
langchain-groq
python-dotenv
langchain_community
GoogleTranslator